+++ pkgs/hooks/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.3
+
+- Bump `package:code_assets` dev_dependency to `^2.0.0`.
+
 ## 2.0.2
 
 - Update documentation for user-defines.