+++ pkgs/objective_c/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 9.4.2
+
+- Support custom target OS and architectures in build hook.
+- Bump `package:code_assets` dependency to `^2.0.0`.
+
 ## 9.4.1
 
 - Fix a [bug](https://github.com/flutter/flutter/issues/186794) related to